Assign a Block* to data instead of reading every byte
Removed the loop that reads each byte individually into a Block
structure. Created a new variable blocks[] which is simply a pointer to
the data. Since blocks[] is declared as a Block*, we can take advantage
of the structure provided by struct Block without having to wastefully
and uselessly copy the data byte by byte to a temporary Block buffer.