Changed strcpy() to strncpy() in parser.c
This is much safer. Previously, strcpy kept copying until it hit the
first null byte it encountered, which was usually the first byte of the
data section - if the first byte of the data section was not null it
would keep copying until it hit one which could be arbitrarily large
causing buffer overflows. Now strnpy() only ever copies 4 bytes.